♻️ Split and expose states
