+++ lib/wechat_camera_picker.dart
@@ -10,6 +10,14 @@ export 'package:photo_manager/photo_manager.dart';
 export 'src/constants/config.dart';
 export 'src/constants/enums.dart';
 export 'src/constants/type_defs.dart';
+
 export 'src/delegates/camera_picker_text_delegate.dart';
+
+export 'src/states/camera_picker_state.dart';
+export 'src/states/camera_picker_viewer_state.dart';
+
+export 'src/widgets/camera_focus_point.dart';
 export 'src/widgets/camera_picker.dart';
 export 'src/widgets/camera_picker_page_route.dart';
+export 'src/widgets/camera_picker_viewer.dart';
+export 'src/widgets/camera_progress_button.dart';
